+++ FociFit.Models/Domains/Account.cs
@@ -1,0 +1,15 @@
+﻿using FociFit.Models.Requests;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace FociFit.Models.Domains
+{
+    public class Account : AccountUpdateRequest
+    {
+        public DateTime CreatedDate { get; set; }
+        public DateTime ModifiedDate { get; set; }
+    }
+}
