+++ FociFit.Models/Requests/MoviesUpdateRequest.cs
@@ -5,6 +5,6 @@ namespace FociFit.Models.Responses
     public class MoviesUpdateRequest : MoviesCreateRequest
     {
         [Required]
-        public int Id { get; set; }
+        public int? Id { get; set; }
     }
 }
